# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Customer Controller
  description: ""
  termsOfService: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: Joel Silva
    email: joeltadeu@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
  - url: http://localhost:9081
    description: Generated server url
...
paths:
...
  /v1/customers/{customerId}/addresses/{id}:
    get:
      tags:
        - Address Controller
      summary: Retrieve a customer address by id
      description: This endpoint is responsible to retrieve the customer address data
        by id
      operationId: findAddressById
      parameters:
        - name: customerId
          in: path
          description: Id of the customer
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: id
          in: path
          description: Id of the address to be searched
          required: true
          schema:
            type: integer
            format: int64
          example: 15
      responses:
        "200":
          description: Return address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressResponse'
        "500":
          description: An unexpected error occurred during retrieve the address
          content:
            application/json:
              examples:
                Internal server error example:
                  description: A internal server error response example when trying
                    to retrieve a customer address
                  value:
                    code: 500
                    status: Internal Server Error
                    description: An error occurred while processing your request
                    date: 2022-10-28T21:39:54.603263862
        "404":
          description: Address not found
          content:
            application/json:
              examples:
                Not found example:
                  description: A not found response example when trying to retrieve
                    a address does not exist
                  value:
                    code: 404
                    status: Not Found
                    description: Address with id 2 was not found
                    date: 2022-10-28T21:39:54.603263862
    put:
      tags:
        - Address Controller
      summary: Update an address
      description: This endpoint is responsible to update a customer's address data
        by id
      operationId: updateAddress
      parameters:
        - name: customerId
          in: path
          description: Id of the customer
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: id
          in: path
          description: Id of the address to be updated
          required: true
          schema:
            type: integer
            format: int64
          example: 15
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
        required: true
      responses:
        "400":
          description: Address request is invalid
          content:
            application/json:
              examples:
                Bad request example:
                  description: A bad request response example when trying to update
                    an address
                  value: |
                    {
                         "code": 400,
                         "status": "Bad Request",
                         "description": "Validation Exception",
                         "date": "2022-11-25T20:14:04.078313400",
                         "attributes": [
                             {
                                 "attribute": "street",
                                 "errors": [
                                     "Street cannot be null"
                                 ]
                             },
                             {
                                 "attribute": "eircode",
                                 "errors": [
                                     Eircode cannot be null"
                                 ]
                             },
                             {
                                 "attribute": "city",
                                 "errors": [
                                     "City cannot be null"
                                 ]
                             },
                             {
                                 "attribute": "country",
                                 "errors": [
                                     "Country cannot be null"
                                 ]
                             }
                         ]
                    }
        "500":
          description: An unexpected error occurred during update the address
          content:
            application/json:
              examples:
                Internal server error example:
                  description: A internal server error response example when trying
                    to update an address
                  value:
                    code: 500
                    status: Internal Server Error
                    description: An error occurred while processing your request
                    date: 2022-10-28T21:39:54.603263862
        "200":
          description: Address updated
          content:
            application/json: {}
        "404":
          description: Address not found
          content:
            application/json:
              examples:
                Not found example:
                  description: A not found response example when trying to update
                    an address does not exist
                  value:
                    code: 404
                    status: Not Found
                    description: Address with id 2 was not found
                    date: 2022-10-28T21:39:54.603263862
    delete:
      tags:
        - Address Controller
      summary: Delete the customer address by id
      description: This endpoint is responsible to delete the customer address by
        id
      operationId: deleteAddress
      parameters:
        - name: customerId
          in: path
          description: Id of the customer
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: id
          in: path
          description: Id of the address to be deleted
          required: true
          schema:
            type: integer
            format: int64
          example: 15
      responses:
        "500":
          description: An unexpected error occurred during delete the customer address
          content:
            application/json:
              examples:
                Internal server error example:
                  description: A internal server error response example when trying
                    to delete customer address
                  value:
                    code: 500
                    status: Internal Server Error
                    description: An error occurred while processing your request
                    date: 2022-10-28T21:39:54.603263862
        "200":
          description: Address deleted
          content:
            application/json: {}
        "404":
          description: Address not found
          content:
            application/json:
              examples:
                Not found example:
                  description: A not found response example when trying to delete
                    a customer address does not exist
                  value:
                    code: 404
                    status: Not Found
                    description: Address with id 2 was not found
                    date: 2022-10-28T21:39:54.603263862
...
  /v1/customers/{customerId}/addresses:
    get:
      tags:
        - Address Controller
      summary: Retrieve a list of addresses by customer id
      description: This endpoint is responsible to retrieve a list of customer addresses
        based on the applied filters
      operationId: listAllAddresses
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: filter
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/AddressFilter'
      responses:
        "200":
          description: Return customer addresses list
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageAddressResponse'
        "500":
          description: An unexpected error occurred during retrieve the customer addresses
            list
          content:
            application/json:
              examples:
                Internal server error example:
                  description: A internal server error response example when trying
                    to retrieve a customer addresses list
                  value:
                    code: 500
                    status: Internal Server Error
                    description: An error occurred while processing your request
                    date: 2022-10-28T21:39:54.603263862
    post:
      tags:
        - Address Controller
      summary: Create an address
      description: This endpoint is responsible to create a new address
      operationId: createAddress
      parameters:
        - name: customerId
          in: path
          description: Id of the customer
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
        required: true
      responses:
        "400":
          description: Address request is invalid
          content:
            application/json:
              examples:
                Bad request example:
                  description: A bad request response example when trying to create
                    an address
                  value: |
                    {
                         "code": 400,
                         "status": "Bad Request",
                         "description": "Validation Exception",
                         "date": "2022-11-25T20:14:04.078313400",
                         "attributes": [
                             {
                                 "attribute": "street",
                                 "errors": [
                                     "Street cannot be null"
                                 ]
                             },
                             {
                                 "attribute": "eircode",
                                 "errors": [
                                     Eircode cannot be null"
                                 ]
                             },
                             {
                                 "attribute": "city",
                                 "errors": [
                                     "City cannot be null"
                                 ]
                             },
                             {
                                 "attribute": "country",
                                 "errors": [
                                     "Country cannot be null"
                                 ]
                             }
                         ]
                    }
        "201":
          description: Address created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressResponse'
        "500":
          description: An unexpected error occurred during creating the address
          content:
            application/json:
              examples:
                Internal server error example:
                  description: A internal server error response example when trying
                    to create an address
                  value:
                    code: 500
                    status: Internal Server Error
                    description: An error occurred while processing your request
                    date: 2022-10-28T21:39:54.603263862
components:
  schemas:
...
    AddressRequest:
      required:
        - city
        - country
        - county
        - eircode
        - street
      type: object
      properties:
        street:
          type: string
          description: Address street
          example: Coosan Road
        complement:
          type: string
          description: Address complement
          example: "Apt. 104, Block 4"
        eircode:
          type: string
          description: Address eircode
          example: N46RP87
        city:
          type: string
          description: Address city
          example: Athlone
        county:
          type: string
          description: Address county
          example: Co. Westmeath
        country:
          type: string
          description: Address country
          example: Ireland
    AddressResponse:
      type: object
      properties:
        id:
          type: integer
          description: Address Id
          format: int64
          example: 13
        street:
          type: string
          description: Address street
          example: Coosan Road
        complement:
          type: string
          description: Address complement
          example: "Apt. 104, Block 4"
        eircode:
          type: string
          description: Address eircode
          example: N46RP87
        city:
          type: string
          description: Address city
          example: Athlone
        county:
          type: string
          description: Address county
          example: Co. Westmeath
        country:
          type: string
          description: Address country
          example: Ireland
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        paged:
          type: boolean
        unpaged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        unsorted:
          type: boolean
        sorted:
          type: boolean
...
    AddressFilter:
      required:
        - pageNumber
        - pageSize
      type: object
      properties:
        pageNumber:
          type: integer
          description: Zero-base page index (0..N)
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: Number of records per page
          format: int32
          example: 10
          default: 10
        sortDirection:
          type: string
          description: Sort direction
          example: ASC|DESC
          enum:
            - ASC
            - DESC
        sortBy:
          type: string
          description: Sort column name
          example: firstName
        street:
          type: string
          description: Street to be searched
          example: Coosan Road
        city:
          type: string
          description: City to be searched
          example: Athlone
        county:
          type: string
          description: County to be searched
          example: Co. Westmeath
        country:
          type: string
          description: Country to be searched
          example: Ireland
        eircode:
          type: string
          description: Eircode to be searched
          example: N47RP97
    PageAddressResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/AddressResponse'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean